lib/alloc_phys: Remove locking
powerpc and s390 do not use the physical allocator.

arm and arm64 initialize the physical allocator in the setup code, only to
drain it immediately afterwards to initialize the page allocator.

x86 calls setup_vm() before any test that allocates memory, which similarly
drains the physical allocator to initialize the page allocator.

The setup code runs on a single core, which means that there is no need to
protect the internal data structures of the physical allocator against
concurrent accesses. Simplify the allocator by removing the locking.

Signed-off-by: Alexandru Elisei <alexandru.elisei@arm.com>