arm/arm64: Map the UART when creating the translation tables
The MMU is now enabled before the UART is probed, which leaves
kvm-unit-tests with a window where attempting to write to the console
results in an infinite data abort loop triggered by the exception
handlers themselves trying to use the console. Get around this by
mapping the UART early address when creating the translation tables.

Signed-off-by: Alexandru Elisei <alexandru.elisei@arm.com>